finished adding the apis, need to test them all still
